backend/requirements.txt
# Core web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Chemistry libraries
rdkit==2023.9.4
numpy==1.26.3
pandas==2.1.4
scipy==1.12.0

# AI/ML
anthropic==0.18.1
scikit-learn==1.4.0
torch==2.1.2

# Molecular docking (AutoDock Vina)
vina==1.2.5
meeko==0.5.0

# Structure prediction
biopython==1.83
py3Dmol==2.0.4

# Database and search
requests==2.31.0
aiohttp==3.9.1
chembl_webresource_client==0.10.8
pubchempy==1.0.4

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
